genai_requirements_old.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
audio-recorder-streamlit==0.0.10
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.5.2
boto3==1.34.162
botocore==1.34.162
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docutils==0.16
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.112.1
filelock==3.15.4
flatbuffers==24.3.25
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
gTTS==2.5.4
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==30.1.0
langchain==0.2.15
langchain-aws==0.1.17
langchain-cli==0.0.30
langchain-community==0.2.15
langchain-core==0.3.74
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langserve==0.2.2
langsmith==0.4.14
libcst==1.4.0
llama-index-core==0.11.1
llama-index-embeddings-bedrock==0.3.0
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.44.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
playwright==1.46.0
posthog==3.6.0
prompt_toolkit==3.0.48
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.2
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
streamlit-bokeh-events==0.1.2
streamlit-sortables==0.3.1
sympy==1.13.2
tavily-python==0.4.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
wheel==0.43.0
wrapt==1.16.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.9.4
zipp==3.20.2
zstandard==0.23.0
# python 3.12.9 used